FLP-based filters
- Switch using new "keep" api on FLP filters
- Support sampling
- Support regexes
- Add tests
