+++ config/crd/bases/flows.netobserv.io_flowcollectors.yaml
@@ -1646,6 +1646,11 @@ spec:
                               - Metrics
                               - Exporters
                             type: string
+                          sampling:
+                            description: '`sampling` is an optional sampling rate to apply to this filter.'
+                            format: int32
+                            minimum: 0
+                            type: integer
                         type: object
                       type: array
                     healthPort:
@@ -7339,6 +7344,11 @@ spec:
                               - Metrics
                               - Exporters
                             type: string
+                          sampling:
+                            description: '`sampling` is an optional sampling rate to apply to this filter.'
+                            format: int32
+                            minimum: 0
+                            type: integer
                         type: object
                       type: array
                     imagePullPolicy:
